Archive initial commit timestamp based on archive
When generating the initial commit for an archive-based repo, we
now try to base it on the time of the archive.  In this way, even
if you generate the git repo from the archive multiple times, it
hopefully has the same hash.  So if, e.g., it's done by multiple
users independently, the histories start in the same place.