Always do initial fetch with --no-tags
We now always do the initial fetch with --no-tags.  Unless the no_tags
config key is set, we then do a new "normal" fetch in order to get
the normal automatic tag behavior.  (This is followed up with the
fetch that gets *all* remote tags and puts them in the special
"remote_tags" namespace.)

This is sort of a hack, but it's really effective.  The purpose is so
that for the initial fetch, we can display the output.  This is
important because it may take a long time and we want to see
progress, and it also shows the branches which is nice.  But since we
now do this with --no-tags, we don't show the list of tags, which is
the point.  They're usually not interesting and it's pretty common
for there to be a large number, so this was pretty spammy.  Now
we fetch them independently, which should be fast, so there's no
big harm in hiding their fetching.