Add Git.run_checkout()
This is a special run method meant for doing checkouts.  In normal
usage, it runs the command with -q and --progress, so we get
progress display (if any), but not a whole lot else.  This is
good because otherwise it often prints out some messages that
aren't especially interesting and/or are largely overlapping
information that's redundant with our log messages.